tweak: DevEx to run changelog independently
